ml-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
scikit-learn==1.3.2
pandas==2.1.3
numpy==1.25.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
redis==5.0.1
structlog==23.2.0
prometheus-client==0.19.0
python-multipart==0.0.6